refactor: create useCookMode, useRecipeShare and useTriedRecipe hooks. Add toggleInset to eliminate duplicate toggle logic in recipe-detail
